added analytics and services file
