+++ package.json
@@ -22,6 +22,7 @@
     "postcss-focus-visible": "^6.0.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "@vercel/analytics": "^1.5.0",
     "sanitize-html": "^2.11.0",
     "tailwindcss": "^3.3.3",
     "typescript": "^5.1.6"
